+++ src/App.js
@@ -4,12 +4,14 @@ import Routes from "./components/Routes";
 import "./app.scss";
 import "./config/firebase.config";
 import "./global.scss";
+// import RegisterAdopter1 from "./pages/cadastro-do-doador/cadastro-tipo-de-pessoa/adopter-registration/register-adopter-data";
 
 
 function App() {
   return (
     <div className="App">
       <Routes />
+      {/* <RegisterAdopter1 /> */}
     </div>
   );
 }
