+++ src/components/Routes/index.jsx
@@ -5,11 +5,11 @@ import CardPet from '../../pages/card-pet';
 import ProfilePerson from '../../pages/perfil-adotante';
 import RegisterPet from '../../pages/register-pet';
 import RegisterAdopter from '../../pages/cadastro-do-doador/cadastro-tipo-de-pessoa';
-import DonorRegister from '../../pages/cadastro-do-doador/cadastro-tipo-de-pessoa';
+import DonorRegister from '../../pages/cadastro-do-doador/cadastro-dados-da-pessoa';
 import ContactDetails from '../../pages/cadastro-do-doador/cadastro-dados-contato';
 import ProfilePet from '../../pages/perfil-pet';
 import Login from '../../pages/login';
-
+import RegisterAdopterData from '../../pages/cadastro-do-doador/cadastro-tipo-de-pessoa/adopter-registration/register-adopter-data'
 export default function Routes() {
 	return (
 		<BrowserRouter>
@@ -19,6 +19,8 @@ export default function Routes() {
 				<Route exact path="/filtro" component={Filter} />
 				<Route exact path="/card-do-Pet" component={CardPet} />
 				<Route exact path="/perfil-do-adotante" component={ProfilePerson} />
+				<Route exact path="/cadastro-dados-adotante" component={RegisterAdopterData} />
+
 				<Route
 					exact
 					path="/cadastro-tipo-de-pessoa"
