+++ src/pages/cadastro-do-doador/cadastro-tipo-de-pessoa/index.jsx
@@ -17,7 +17,7 @@ export default function RegisterAdopter() {
     function chooseScreen() {
         console.log(person);
         const location = {
-            pathname: person === "juridica" ? "/cadastro-pessoa-juridica" : "/cadastro-pessoa-fisica",
+            pathname: person === "adotante" ? "/cadastro-dados-adotante" : "/cadastro-do-doador",
             state: {
                 person
             }
@@ -28,16 +28,16 @@ export default function RegisterAdopter() {
 
     return (
         <>
-            <Header title="Você quer doar!" subTitle="Vamos lá!" />
+            <Header title="Vamos nos cadastar!" subTitle="Escolha a forma de cadastro." />
             <Text type="titleform">Vamos começar!</Text>
             <section className="contentform">
                 <div className="form">
                     <div className="textform">
-                        <Text type="labelform">Você é:</Text>
+                        <Text type="labelform">Deseja se cadastrar como :</Text>
                     </div>
                     <div className="form-box">
-                        <RadioButton text="Pessoa Física" name="typePerson" id="fisica" onChange={setPerson} />
-			            <RadioButton text="Pessoa Jurídica (ONG)" name="typePerson" id="juridica" onChange={setPerson} />
+                        <RadioButton text="Adotante" name="typePerson" id="adotante" onChange={setPerson} />
+			            <RadioButton text="Doador" name="typePerson" id="doador" onChange={setPerson} />
                     </div>
                 </div>
             </section>
