+++ src/pages/cadastro-do-doador/cadastro-tipo-de-pessoa/styles.scss
@@ -8,6 +8,9 @@
   height: 83px;
   background: linear-gradient(91.99deg, #008080 45.19%, #800080 110.51%);
   box-shadow: 0px 5px 8px rgba(0, 0, 0, 0.15);
+  .subtitle-header {
+    margin-top: 5px;
+  }
 }
 
 .textheader {
